SCU/requirements.txt
numpy
Pillow
easyocr
requests
PyExecJS
